refactor: Move detect-grammar.ts & validate.ts under extra/
