+++ packages/messageformat/src/extra/detect-grammar.ts
