+++ packages/messageformat/src/extra/validate.ts
@@ -3,10 +3,10 @@ import {
   MessageGroup,
   Resource,
   PatternElement
-} from './data-model';
-import { isFunction } from './pattern/function';
-import { isTerm } from './pattern/term';
-import type { Runtime } from './runtime';
+} from '../data-model';
+import { isFunction } from '../pattern/function';
+import { isTerm } from '../pattern/term';
+import type { Runtime } from '../runtime';
 
 export function validate(
   resources: Iterable<Readonly<Resource>>,
