+++ packages/messageformat/src/index.ts
@@ -5,4 +5,4 @@ export { MessageFormat } from './messageformat';
 export * from './pattern';
 export { ResourceReader } from './resource-reader';
 export * from './runtime';
-export { validate } from './validate';
+export { validate } from './extra/validate'; // must be after ./messageformat -- but why!?
